Add background timers (app.every/app.daily), resolve all audit findings
Background work: implement app.every(ms, fn) and app.daily("HH:MM", fn)
for both Lua and JS runtimes. Timer callbacks run on the event loop
thread with full async support (hull.sleep, http.fetch, db.*). Adds
detached async mode for connectionless operations, self-cancellation
via return false, and one-invocation-at-a-time guards.

New example: examples/timers/ with E2E tests for both runtimes (188/188
e2e example tests pass). Timers example demonstrates repeating heartbeat
timer and self-cancelling counter.

Security audit: consolidated C/Lua/JS audit into docs/audit_report_20260317.md.
All 45 findings resolved (38 fixed, 7 documented as acceptable):
- FTS5 tokenize injection (only unparameterized SQL in stdlib)
- Cookie name/value/domain validation (header injection prevention)
- Session ID format validation (64-char hex gate)
- TOCTOU races in inbox/idempotency (db.batch, INSERT OR IGNORE)
- CORS preflight headers spec compliance
- RBAC 401 vs 403 distinction + configurable user_id extraction
- JWT require_exp option, nbf check, named EXP_RELATIVE_THRESHOLD
- CSRF requireSession option, ASCII secret enforcement
- CSV max_rows, form max_fields limits
- Build hardening: _FORTIFY_SOURCE=2, INSTR_COUNT clamp, atoi->strtol
- email.js named export consistency, verify.js JSON.parse safety
- search.js var->const/let modernization

+++ src/hull/worker_db.c
@@ -120,7 +120,7 @@ static int db_result_grow(HlDbResult *r)
     int new_cap = r->capacity ? r->capacity * 2 : 16;
     if (new_cap > 100000) return -1; /* sanity cap */
     HlDbValue *nv = realloc(r->values,
-                             (size_t)(new_cap * r->ncols) * sizeof(HlDbValue));
+                             (size_t)new_cap * (size_t)r->ncols * sizeof(HlDbValue));
     if (!nv) return -1;
     r->values = nv;
     r->capacity = new_cap;
